修改示例，rename 命名
