+++ components/mip-analytics/data-collection.js
