+++ components/mip-analytics/example/mip-analytics-alexa.html
@@ -5,7 +5,7 @@
   <meta name="viewport" content="width=device-width,minimum-scale=1,initial-scale=1">
   <title>MIP page</title>
   <link rel="canonical" href="https://mip.xxx.com/对应的原页面地址">
-  <link rel="stylesheet" href="http://bos.nj.bpc.baidu.com/v1/assets/mip/projects/mip.css">
+  <link rel="stylesheet" href="https://c.mipcdn.com/static/v2/mip.css">
 </head>
 <body>
   <mip-analytics type="alexametrics">
@@ -18,43 +18,7 @@
       }
     </script>
   </mip-analytics>
-  <script src="http://bos.nj.bpc.baidu.com/v1/assets/mip/projects/mip.js"></script>
+  <script src="https://c.mipcdn.com/static/v2/mip.js"></script>
   <script src="/mip-analytics/mip-analytics.js"></script>
 </body>
 </html>
-
-
-<mip-analytics>
-  <script type="application/json">
-  {
-    "hosts": {
-      "disp": "//certify-amp.alexametrics.com/atrk.gif?atrk=${atrk_acct}&domain=${domain}&"
-    },
-    "setting": {
-      "disp": [
-        {
-          "host": "disp",
-          "queryString": {
-            "jsv": "mip-${mipVersion}",
-            "frame_height": "${viewportHeight}",
-            "frame_width": "${viewportWidth}",
-            "title": "${title}",
-            "time": "${timestamp}",
-            "time_zone_offset": "${timezone}",
-            "screen_params": "${screenWidth}x${screenHeight}x${screenColorDepth}",
-            "ref_url": "${documentReferrer}",
-            "host_url": "${sourceUrl}",
-            "random_number": "${random}",
-            "user_lang": "${browserLanguage}",
-            "mip_doc_url": "${mipdocUrl}"
-          },
-          "vars": {
-            "atrk_acct": "xxxx",
-            "domain": "xxx.oo"
-          }
-        }
-      ]
-    }
-  }
-  </script>
-</mip-analytics>
