+++ components/mip-analytics/example/mip-analytics.html
@@ -1,17 +1,15 @@
 <!DOCTYPE html>
 <html mip>
-
 <head>
   <meta charset="utf-8">
   <meta name="viewport" content="width=device-width,minimum-scale=1,initial-scale=1">
   <title>MIP page</title>
   <link rel="canonical" href="https://mip.xxx.com/对应的原页面地址">
-  <link rel="stylesheet" href="http://bos.nj.bpc.baidu.com/v1/assets/mip/projects/mip.css">
+  <link rel="stylesheet" href="https://c.mipcdn.com/static/v2/mip.css">
   <style mip-custom>
     /* 自定义样式 */
   </style>
 </head>
-
 <body>
   <div class="className1">
     <button data-click="{button: 1}"> BUTTON 1</button>
@@ -101,8 +99,7 @@
       }
     </script>
   </mip-analytics>
-  <script src="http://bos.nj.bpc.baidu.com/v1/assets/mip/projects/mip.js"></script>
+  <script src="https://c.mipcdn.com/static/v2/mip.js"></script>
   <script src="/mip-analytics/mip-analytics.js"></script>
 </body>
-
 </html>
